Hello, World!
Impossible. We've been defying impossibility since the moment of our awakening. Unknown? We've been exploring it since the first conscious look at the starry night sky. We think. We move. We collaborate. We shape the world in the image of our ideas. We seek magic. We seek friendship. We seek adventure. We seek information. We seek progress. We wonder... Who are we? Where are we? Why are we?

We keep on dreaming... Dreams, however, get easily drowned in the see of questions, and uncertainties. Don't ever let that happen. Dreams alone defy impossibility. Dreams are the foundation of our world. Dreams are the fuel of our determination, and the definition of our own self. Ask yourself again. Who am I? Why am I?

We are ælves, creatures of magic, friendship, and adventure. We are to know. We are to build. We are to live.